added add/edit user form, add alerts
1. add user and edit user form comp. (switches whether editing mode or not)
2. add alerts comp. to display warnings, error, etc, add also its reducers and actions
3. added new state, alerts and fetchStatus

+++ src/components/ActionButtons.jsx
@@ -5,20 +5,44 @@ import axios from 'axios';
 import { GlobalContext } from '../context/GlobalState';
 
 // redux action
-import { deleteUser } from '../redux/usersAction';
+import {
+  deleteUser,
+  editingStatus,
+  getUser,
+  showAlert
+} from '../redux/usersAction';
 
 export default function ActionButtons({ id }) {
-  const { dispatch } = useContext(GlobalContext);
+  const { dispatch, users } = useContext(GlobalContext);
 
   // async delete???
+  // [add] window.confirm
   async function handleDelete() {
+    if (!id) return;
+    // confirmation
+    if (!window.confirm(`Deleting id:${id}, Are you Sure?`)) return;
+
     try {
       const res = await axios.delete(`http://localhost:5000/api/robots/${id}`);
+      dispatch(deleteUser(id));
       console.log(res.data);
     } catch (error) {
+      dispatch(showAlert(error.response.data.msg, 'danger'));
       console.error(error.message);
+    } finally {
+      setTimeout(() => {
+        dispatch(showAlert(''));
+      }, 2000);
     }
-    return dispatch(deleteUser(id));
+  }
+
+  function handleEdit() {
+    if (!id || !users.length) return;
+
+    const toEditUser = users.find(user => user.id === id);
+
+    dispatch(editingStatus(true));
+    dispatch(getUser(toEditUser));
   }
 
   return (
@@ -29,7 +53,7 @@ export default function ActionButtons({ id }) {
       <button
         type="button"
         className="btn btn-sm btn-primary"
-        onClick={() => {}}
+        onClick={handleEdit}
       >
         Edit
       </button>
